requirements.txt
alembic==1.8.1
apache-airflow-backport-providers-amazon==2021.3.3
apache-airflow==1.10.12
apispec==1.3.3
argcomplete==1.12.3
attrs==19.3.0
Babel==2.10.3
boto3==1.15.18
botocore==1.18.18
cached-property==1.5.2
cattrs==1.0.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.37
defusedxml==0.7.1
dill==0.3.5.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==0.14.3
Flask==1.1.4
funcsigs==1.0.2
future==0.18.2
graphviz==0.20.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.1
Markdown==2.6.11
MarkupSafe==2.1.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
marshmallow==2.21.0
natsort==8.1.0
numpy==1.21.6
pandas==1.1.5
pendulum==1.4.4
pip==21.3.1
prison==0.2.1
psutil==5.9.1
Pygments==2.12.0
PyJWT==1.7.1
pyrsistent==0.18.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.3.7
setproctitle==1.3.0
setuptools==60.2.0
six==1.16.0
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.38.3
SQLAlchemy==1.4.39
tabulate==0.8.10
tenacity==4.12.0
text-unidecode==1.3
thrift==0.16.0
typing-extensions==4.3.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
watchtower==0.7.3
Werkzeug==0.16.1
wheel==0.37.1
WTForms==3.0.1
zipp==3.8.1
zope.deprecation==4.4.0
